refactor(create-access-token): move it into utils
