+++ src/backend/index.ts
@@ -4,7 +4,7 @@ import path from 'path'
 import { v4 as uuid_v4 } from 'uuid'
 import * as t from 'io-ts'
 import { NextFunction, Request, Response } from 'express'
-import { createAccessToken } from './create-acccess-token'
+import { createAccessToken } from './util/create-acccess-token'
 import { registerLtiPlatforms } from './register-lti-platforms'
 import urlJoin from 'url-join'
 import config from '../utils/config'
